work around UI glitch in "light mode" ... by going back to a slightly less annoying (at least to me) UI glitch
There was a "hack" in place to fix the UI issue in light mode where you don't see the selection in the command text field. BUT, this actually introduced a worse glitch that wasn't immediately obvious: when the terminal output "grows" while the window isn't visible, and if you then reactivate the (now taller) window, the new window part will be missing the expected blur, and will instead be mostly transparent (with the output "visible").

So for now we deactivate that hack ... and live with the "invisible" selection highlight color.

If anyone has a solution that fixes both of these at the same time, do let me know.